+++ example/lib/models/calendar_color_model.dart
@@ -3,6 +3,6 @@ import 'package:flutter/material.dart';
 class CalendarColorModel{
   String title;
   Color  color;
-
-  CalendarColorModel({@required this.title,@required this.color});
+   Widget body;
+  CalendarColorModel({@required this.title,@required this.color,@required this.body});
 }
